==> 01-- Data fetching.ipynb <==
#import necessary library
import pandas as pd
import requests

#this function return the dataframe with the text data after getting it from api
def get_text_featuer(data):
    data['id'] = data['id'].apply(str)
    data['text'] = ""
    i = 0
    #we will iterate over all records with every 1000 record because it's the max number of values can api take
    while i < data.shape[0]: 
        #this if condition for incase the last remain records less than 1000 so just take it as itis and
        # break or finish loop aftre that
        if data.shape[0] - i < 1000:
            print(data.shape[0] - i)
            ids = data.iloc[i:]['id'].tolist()
            resp = requests.post(url = 'https://recruitment.aimtechnologies.co/ai-tasks', json = ids)
            json_rec = resp.json()
            data['text'].iloc[i:] = list(json_rec.values())
            break
        #to follow number of ids that already getten their texts
        print(i)
        ids = data.iloc[i:i+1000]['id'].tolist()
        resp = requests.post(url = 'https://recruitment.aimtechnologies.co/ai-tasks', json = ids)
        json_rec = resp.json()
        data['text'].iloc[i:i+1000] = list(json_rec.values())
        
        i += 1000
    return data    
    

#read the dataframe and apply the function to get the text of ids
df = pd.read_csv('dialect_dataset.csv')
texted_data = get_text_featuer(df)

#show some of data and to make sure that get the right text take some of this id's and send request to api using postman
#app and compare the result in postman with the results here
pd.set_option('display.max_colwidth', 100000)
pd.set_option("display.min_rows", 30)
texted_data.tail(30)

#Export the new data
texted_data.to_csv(r'texted_dialect_dataset.csv', index = False)
