==> 02-- Data pre-processing .ipynb <==
#import nessery libraries
import pandas as pd
import re

#inspect our data 
pd.set_option('display.max_colwidth', 100000)
pd.set_option("display.min_rows", 20)
text_data = pd.read_csv('texted_dialect_dataset.csv')
text_data.head()

x = text_data.drop('dialect')
y = text_data['dialect']
x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=.2, stratify=y, random_state=42)

## Data Cleaning

# all we care about in our text data is the arabics word because it's can specify the dielect according to it only so we  
# will erase all other caracters like english words symbols and emojis
text_data['text'] = text_data['text'].apply(lambda x:re.sub(r"([@A-Za-z0-9_ـــــــــــــ]+)|[^\w\s]|#|http\S+", "", x))

#to remove tleading and trailing whitespaces in strings in our data espiacially spaces preduced from pervious steps
text_data['text'] = text_data['text'].apply(lambda x:x.strip())

#inspect data after clean
text_data.head(20)

text_data.isnull().sum(axis = 0)

# after clear the text from all unwanted caracters it's may preduce some nan values in text column so we need to clean it
text_data.dropna(axis=0,inplace=True)

text_data.isna().sum()

**Notes :** 
- we don't need to apply stemming or lemmtization in our texts because there are unique words or combination of words that may help to classify dialect of the text easily so stemming and Lemmatization may destroy this context, expersions or words
- For word embedding we will perform it in model because we can take many approches with different models according to logic we are follow

#Export the new data
text_data.to_csv(r'text_ArabDialect_dataset.csv', index = False)
