==> 03_Model_Training.ipynb <==
#import nessery libraries
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.model_selection import GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.svm import LinearSVC
from sklearn.linear_model import SGDClassifier
from sklearn.pipeline import make_pipeline
from sklearn.pipeline import Pipeline
from sklearn import metrics
import time
import tensorflow as tf
import spacy

#inspect our data 
pd.set_option('display.max_colwidth', 100000)
pd.set_option("display.min_rows", 20)
text_data = pd.read_csv('text_ArabDialect_dataset.csv')
text_data.head()

text_data.isna().sum()

text_data[text_data['text'].isna() == True]

#because the jupyter notebook not read any nan values so I will erase this nan here in colab in this file not in preprosessing file
text_data.dropna(axis=0,inplace=True)

text_data.isna().sum()

x = text_data['text']
y = text_data['dialect']
x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=.2, stratify=y, random_state=42)

x_train.head()

## Classification Models
- SVM classifier when using word n-grams
- SGD classifier when using a Tf-idf Vectorizer
- RNN with contextualized embeddings

#intiate the function that apply different accurcy metric on prdicted data 
def print_report(pipe, x_test, y_test):
    y_pred = pipe.predict(x_test)
    report = metrics.classification_report(y_test, y_pred)
    print(report)
    print("accuracy: {:0.3f}".format(metrics.accuracy_score(y_test, y_pred)))

### SVM classifier when using a combination of character and word n-grams
**The logic behined this approche is that there is uniqe words and expersion that can identify delict easily so using word n-gram will help model to focus in this important featuer**

#intiate word-ngram BOW for embedding and svm for training model and fit them in pipeline 
# also calculate the time that model need to fit for comparison porpose 

# starting time
start = time.time()
vec = CountVectorizer(analyzer='word', ngram_range=(1, 3))
clf = LinearSVC(max_iter=1500)
svm_wordNgram_pipe = make_pipeline(vec, clf)
svm_wordNgram_pipe.fit(x_train, y_train)

# end time
end = time.time()

# total time taken
print(f"Runtime of the Model to fit {end - start}")

print_report(svm_wordNgram_pipe, x_test, y_test)

import joblib
joblib.dump(svm_wordNgram_pipe, 'svm_wordNgram_pipe.pkl')

### SGD classifier when using a Tf-idf Vectorizer
**The logic behined this approche is making th unique words that identify delict easily and have less repetation so we need to make this words or expersion have more weight than others so tf-idf will be very good for it also for data training model sgd is the stokastic gradiant desent it's an optomization function used to reach to optimal is stokastic way but here in the implementation in sklearn it can implement loss function of linear calssification model like linearSVM and also this model is very good in working with large-scale and sparse problem like our problem so it solve the converge problem in pervious linearSVM model...also applaied it in grid search to tune the hyperparameters like max_df that specifiy the word considration until specific frequancy and max featuers that you can take**

#just for getting the number of combination in our gridsearch to estimate if it will take very long time or not
from sklearn.model_selection import ParameterGrid
parameters = {'tfidf__ngram_range': [(1, 2),(1, 3)],
              'tfidf__max_features': [500, 1000, None],
              'tfidf__norm': ('l1', 'l2'),
              }
gd = ParameterGrid(parameters)
len(gd)

# starting time
start = time.time()

sgd_wordNgram_pipe = Pipeline([
  ('tfidf', TfidfVectorizer()),
  ('sgd', SGDClassifier())
  ])

# sgd_wordNgram_pipe = make_pipeline(tfidf, sgd)
parameters = {'tfidf__ngram_range': [(1, 2),(1, 3)],
              'tfidf__max_features': [500, 1000, None],
              'tfidf__norm': ('l1', 'l2'),
              }

grid = GridSearchCV(sgd_wordNgram_pipe, parameters, cv=2, verbose=1)
grid.fit(x_train, y_train)
# end time
end = time.time()

# total time taken
print(f"Runtime of the Model to fit {end - start}")

grid.best_estimator_

print_report(grid.best_estimator_, x_test, y_test)

import joblib
joblib.dump(grid.best_estimator_, 'sgd_wordNgram_pipe.pkl')

### RNN with contextualized embeddings

from tensorflow.keras.layers import GRU, Embedding, Dense, Input, Dropout, Bidirectional
from tensorflow.keras.models import Sequential
from keras.preprocessing.text import Tokenizer, text_to_word_sequence
from keras.preprocessing.sequence import pad_sequences
from sklearn.utils import shuffle

tknzr = Tokenizer(lower=True, split=" ")
tknzr.fit_on_texts(x)

#making sequences:
X_rnn = tknzr.texts_to_sequences(x)
X_rnn = pad_sequences(X_rnn, padding='post', value=0)

y_rnn = pd.get_dummies(y)
y_rnn.shape

model = tf.keras.models.Sequential([    
    tf.keras.layers.Embedding(len(tknzr.word_index), 64),
    tf.keras.layers.SimpleRNN(64),
    tf.keras.layers.Dense(64, activation='relu'),
    tf.keras.layers.Dense(18, activation='sigmoid')
])

model.compile(loss='binary_crossentropy',optimizer=tf.keras.optimizers.Adam(),
              metrics=['accuracy'])

model.summary()

model.fit(X_rnn, y_rnn, validation_split = 0.1, epochs = 3, batch_size= 128, shuffle = True)

test_loss, test_acc = model.evaluate(x_test_padded, y_test)

print('Test Loss: {}'.format(test_loss))
print('Test Accuracy: {}'.format(test_acc))
